requirements.txt
ackage                       Version
----------------------------- -----------
altgraph                      0.17.3
appdirs                       1.4.4
argon2-cffi                   21.3.0
argon2-cffi-bindings          21.2.0
asgiref                       3.5.2
asttokens                     2.0.8
attrs                         22.1.0
backcall                      0.2.0
backports.functools-lru-cache 1.6.6
beautifulsoup4                4.11.1
biopython                     1.81
bleach                        5.0.1
blinker                       1.6.2
Brotli                        1.1.0
cffi                          1.15.1
charset-normalizer            3.2.0
click                         8.1.6
colorama                      0.4.5
contourpy                     1.0.5
cssselect2                    0.7.0
cycler                        0.11.0
debugpy                       1.6.3
decorator                     5.1.1
defusedxml                    0.7.1
Django                        4.1.2
dna-features-viewer           3.1.2
dnacauldron                   2.0.9
dnachisel                     3.2.11
docopt                        0.6.2
entrypoints                   0.4
et-xmlfile                    1.1.0
executing                     1.1.1
fastjsonschema                2.16.2
flametree                     0.1.12
Flask                         2.3.2
fonttools                     4.37.4
fuzzywuzzy                    0.18.0
genedom                       0.1.11
html2text                     2020.1.16
html5lib                      1.1
ipykernel                     6.16.0
ipython                       8.5.0
ipython-genutils              0.2.0
ipywidgets                    8.0.4
itsdangerous                  2.1.2
jedi                          0.18.1
Jinja2                        3.1.2
joblib                        1.2.0
jsonschema                    4.16.0
jupyter                       1.0.0
jupyter_client                7.3.5
jupyter-console               6.4.4
jupyter-core                  4.11.1
jupyterlab-pygments           0.2.2
jupyterlab-widgets            3.0.5
kiwisolver                    1.4.4
Levenshtein                   0.21.1
Markdown                      3.4.4
MarkupSafe                    2.1.1
matplotlib                    3.6.3
matplotlib-inline             0.1.6
mistune                       2.0.4
nbclient                      0.7.0
nbconvert                     7.2.1
nbformat                      5.6.1
nest-asyncio                  1.5.6
networkx                      3.1
nodejs                        0.1.1
notebook                      6.4.12
numpy                         1.23.3
openpyxl                      3.1.2
optional-django               0.1.0
packaging                     21.3
pandas                        1.5.0
pandocfilters                 1.5.0
parso                         0.8.3
patsy                         0.5.3
pdf-reports                   0.3.5
pefile                        2023.2.7
pickleshare                   0.7.5
Pillow                        9.2.0
prettytable                   3.8.0
primer3-py                    2.0.0
primerserver2                 2.0.0b19
proglog                       0.1.10
progressbar2                  4.2.0
prometheus-client             0.14.1
prompt-toolkit                3.0.31
psutil                        5.9.2
pure-eval                     0.2.2
py                            1.11.0
pycparser                     2.21
pydenticon                    0.3.1
pydna                         5.2.0
pydyf                         0.7.0
pyfiglet                      0.8.post1
Pygments                      2.13.0
pyinstaller                   6.0.0
pyinstaller-hooks-contrib     2023.8
pyparsing                     3.0.9
pyperclip                     1.8.2
pyphen                        0.14.0
pypugjs                       5.9.12
pyrsistent                    0.18.1
python-box                    7.1.1
python-codon-tables           0.1.12
python-dateutil               2.8.2
python-dotenv                 1.0.0
python-Levenshtein            0.21.1
python-utils                  3.7.0
pytz                          2022.4
pywin32                       304
pywin32-ctypes                0.2.2
pywinpty                      2.0.8
pyzmq                         24.0.1
qtconsole                     5.4.0
QtPy                          2.3.0
rapidfuzz                     3.3.0
reportlab                     4.0.4
scikit-learn                  1.1.2
scipy                         1.9.2
seaborn                       0.12.2
Send2Trash                    1.8.0
sequenticon                   0.1.5
six                           1.16.0
sklearn                       0.0
snapgene-reader               0.1.20
soupsieve                     2.3.2.post1
sqlparse                      0.4.3
stack-data                    0.5.1
statsmodels                   0.13.5
terminado                     0.16.0
threadpoolctl                 3.1.0
tinycss2                      1.1.1
tornado                       6.2
touch                         2020.12.3
tqdm                          4.66.1
traitlets                     5.4.0
typing_extensions             4.7.1
tzdata                        2022.6
values                        2020.12.3
wcwidth                       0.2.5
weasyprint                    59.0
webencodings                  0.5.1
Werkzeug                      2.3.6
widgetsnbextension            4.0.5
xlrd                          2.0.1
xlwt                          1.3.0
xmltodict                     0.13.0
zopfli                        0.2.3